RIP, foolz
The foolz server has been down long enough, it's time to remove it from the defaults.

Add a long-overdue note to the README about configuration, with a pointer to the only
other AniTV server I've been able to find (well, that works) and a disclaimer that it
requires an API key.